# Services/order-service/openapi/watchlists.yaml
openapi: 3.0.3
info:
  title: Watchlists and price alerts
  version: 1.0.0
  description: |
    Sprint 10 module `com.yellow.trade.watchlists`, inside the Trade REST API. Written before
    the controller and brought to the instructor on day one.

    Watchlists a customer names and fills, each entry with the latest price from `market-data`
    (consumer group `watchlist-service`); a watchlist entry is not a position, so anything
    listed can be watched. Price alerts with a threshold and a direction, checked against
    every quote on the stream.

    An ABOVE alert fires on the first quote at or above its threshold, a BELOW alert at or
    below. It fires once and is then TRIGGERED until the customer re-arms it (decision log
    0007). Firing and handing the alert to the notifications module's `AlertDelivery`
    interface are one transaction, so an alert never reads TRIGGERED without a notification
    queued for it (decision log 0008).

    Caps per account: 5 watchlists, 50 entries in each, 20 active alerts. Past a cap the answer
    is `LIM-409`. Every route is the caller's own account only.

    Same error envelope as the contracts: `{ "errorCode", "message" }`.

servers:
  - url: http://localhost:8081
    description: Local stack

security:
  - bearerAuth: []

tags:
  - name: Watchlists
    description: Instruments the customer keeps an eye on.
  - name: Alerts
    description: Price thresholds the customer wants to be told about.

paths:
  /api/v1/accounts/{id}/watchlists:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags: [Watchlists]
      operationId: getWatchlists
      summary: The account's watchlists, in order, each entry with its latest price
      responses:
        '200':
          description: In the customer's order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Watchlist'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    post:
      tags: [Watchlists]
      operationId: createWatchlist
      summary: Create a watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistRequest'
      responses:
        '201':
          description: Created, empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Watchlist'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'

  /api/v1/accounts/{id}/watchlists/{watchlistId}:
    parameters:
      - $ref: '#/components/parameters/AccountId'
      - $ref: '#/components/parameters/WatchlistId'
    put:
      tags: [Watchlists]
      operationId: renameWatchlist
      summary: Rename a watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistRequest'
      responses:
        '200':
          description: Renamed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Watchlist'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Watchlists]
      operationId: deleteWatchlist
      summary: Delete a watchlist and its entries
      responses:
        '204':
          description: Gone
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'

  /api/v1/accounts/{id}/watchlists/{watchlistId}/items/{symbol}:
    parameters:
      - $ref: '#/components/parameters/AccountId'
      - $ref: '#/components/parameters/WatchlistId'
      - $ref: '#/components/parameters/Symbol'
    put:
      tags: [Watchlists]
      operationId: addWatchlistItem
      summary: Add an instrument to the end of a watchlist
      description: Adding one already there changes nothing.
      responses:
        '204':
          description: In the watchlist
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Watchlists]
      operationId: removeWatchlistItem
      summary: Take an instrument off a watchlist
      responses:
        '204':
          description: Not in the watchlist
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'

  /api/v1/accounts/{id}/alerts:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags: [Alerts]
      operationId: getAlerts
      summary: The account's alerts, newest first, in every state
      responses:
        '200':
          description: Newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceAlert'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    post:
      tags: [Alerts]
      operationId: createAlert
      summary: Set a price alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRequest'
      responses:
        '201':
          description: Set, ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceAlert'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'

  /api/v1/accounts/{id}/alerts/{alertId}:
    parameters:
      - $ref: '#/components/parameters/AccountId'
      - $ref: '#/components/parameters/AlertId'
    delete:
      tags: [Alerts]
      operationId: cancelAlert
      summary: Cancel an alert; it stays in the list as CANCELLED
      responses:
        '204':
          description: Cancelled
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'

  /api/v1/accounts/{id}/alerts/{alertId}/rearm:
    parameters:
      - $ref: '#/components/parameters/AccountId'
      - $ref: '#/components/parameters/AlertId'
    post:
      tags: [Alerts]
      operationId: rearmAlert
      summary: Make a triggered or cancelled alert ACTIVE again
      responses:
        '200':
          description: ACTIVE again
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceAlert'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    AccountId:
      name: id
      in: path
      required: true
      description: The numeric account key; must be the token's own account (else `ACC-403`).
      schema:
        type: integer
        format: int64
        minimum: 1
    WatchlistId:
      name: watchlistId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    AlertId:
      name: alertId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Symbol:
      name: symbol
      in: path
      required: true
      description: A listed instrument's symbol, as an order names it.
      schema:
        type: string
        maxLength: 30

  responses:
    Error:
      description: The platform's error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    Watchlist:
      type: object
      required: [id, name, position, items]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        position:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistItem'

    WatchlistItem:
      type: object
      required: [symbol, position, stale]
      properties:
        symbol:
          type: string
        position:
          type: integer
        lastPrice:
          type: number
          nullable: true
          description: The latest price on `market-data`; null until one has arrived.
        changePercent:
          type: number
          nullable: true
        priceAsOf:
          type: string
          format: date-time
          nullable: true
        stale:
          type: boolean

    WatchlistRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 40

    PriceAlert:
      type: object
      required: [id, symbol, direction, threshold, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        symbol:
          type: string
        direction:
          type: string
          enum: [ABOVE, BELOW]
        threshold:
          type: number
        status:
          type: string
          enum: [ACTIVE, TRIGGERED, CANCELLED]
        createdAt:
          type: string
          format: date-time
        triggeredAt:
          type: string
          format: date-time
          nullable: true
        triggeredPrice:
          type: number
          nullable: true
          description: The price on the quote that crossed it.
        notificationId:
          type: string
          format: uuid
          nullable: true
          description: The notification it queued; its delivery is on the notifications inbox.

    AlertRequest:
      type: object
      additionalProperties: false
      required: [symbol, direction, threshold]
      properties:
        symbol:
          type: string
          maxLength: 30
        direction:
          type: string
          enum: [ABOVE, BELOW]
        threshold:
          type: number
          exclusiveMinimum: true
          minimum: 0
          multipleOf: 0.01

    ErrorResponse:
      type: object
      required: [errorCode, message]
      properties:
        errorCode:
          type: string
          description: |
            Two codes added to the catalogue: `LIM-409`, a per-account cap reached, and
            `WCH-404`, no such watchlist or alert on this account.
          enum: [ACC-403, AUTH-401, VAL-422, INS-404, LIM-409, WCH-404]
        message:
          type: string